+++ MapboxNavigation/NavigationView.swift
@@ -44,15 +44,6 @@ open class NavigationView: UIView {
         static let buttonSpacing: CGFloat = 8.0
     }
     
-    lazy var bannerShowConstraints: [NSLayoutConstraint] = [
-        self.instructionsBannerView.topAnchor.constraint(equalTo: self.safeTopAnchor),
-        self.instructionsBannerContentView.topAnchor.constraint(equalTo: self.topAnchor)]
-    
-    lazy var bannerHideConstraints: [NSLayoutConstraint] = [
-        self.informationStackView.bottomAnchor.constraint(equalTo: self.topAnchor),
-        self.instructionsBannerContentView.topAnchor.constraint(equalTo: self.instructionsBannerView.topAnchor)
-    ]
-    
     lazy var endOfRouteShowConstraint: NSLayoutConstraint? = self.endOfRouteView?.bottomAnchor.constraint(equalTo: self.safeBottomAnchor)
     
     lazy var endOfRouteHideConstraint: NSLayoutConstraint? = self.endOfRouteView?.topAnchor.constraint(equalTo: self.bottomAnchor)
@@ -74,16 +65,6 @@ open class NavigationView: UIView {
         return map
     }()
     
-    lazy var instructionsBannerContentView: InstructionsBannerContentView = .forAutoLayout()
-    
-    lazy var instructionsBannerView: InstructionsBannerView = {
-        let banner: InstructionsBannerView = .forAutoLayout()
-        banner.delegate = delegate
-        return banner
-    }()
-    
-    lazy var informationStackView = UIStackView(orientation: .vertical, autoLayout: true)
-    
     lazy var floatingStackView: UIStackView = {
         let stackView = UIStackView(orientation: .vertical, autoLayout: true)
         stackView.distribution = .equalSpacing
@@ -95,14 +76,6 @@ open class NavigationView: UIView {
     lazy var muteButton = FloatingButton.rounded(image: Images.volumeUp, selectedImage: Images.volumeOff)
     lazy var reportButton = FloatingButton.rounded(image: Images.feedback)
     
-    lazy var lanesView: LanesView = .forAutoLayout(hidden: true)
-    lazy var nextBannerView: NextBannerView = .forAutoLayout(hidden: true)
-    lazy var statusView: StatusView = {
-        let view: StatusView = .forAutoLayout()
-        view.isHidden = true
-        return view
-    }()
-    
     lazy var resumeButton: ResumeButton = .forAutoLayout()
     
     lazy var wayNameView: WayNameView = {
@@ -112,7 +85,9 @@ open class NavigationView: UIView {
         return view
     }()
     
-    lazy var bottomBannerContainerView: BottomBannerContainerView = .forAutoLayout()
+    lazy var topBannerContainerView: BannerContainerView = .forAutoLayout()
+    
+    lazy var bottomBannerContainerView: BannerContainerView = .forAutoLayout()
 
     weak var delegate: NavigationViewDelegate? {
         didSet {
@@ -156,58 +131,34 @@ open class NavigationView: UIView {
     }
     
     func setupStackViews() {
-    
-        setupInformationStackView()
         floatingStackView.addArrangedSubviews([overviewButton, muteButton, reportButton])
-    }
-    
-    func setupInformationStackView() {
-        let informationChildren: [UIView] = [instructionsBannerView, lanesView, nextBannerView, statusView]
-        informationStackView.addArrangedSubviews(informationChildren)
-        
-        informationChildren.forEach {
-            $0.leadingAnchor.constraint(equalTo: informationStackView.leadingAnchor).isActive = true
-            $0.trailingAnchor.constraint(equalTo: informationStackView.trailingAnchor).isActive = true
-        }
-    }
-    
-    func setupContainers() {
-        let containers: [(UIView, UIView)] = [
-            (instructionsBannerContentView, instructionsBannerView)
-        ]
-        containers.forEach { $0.addSubview($1) }
     }
     
     func setupViews() {
         setupStackViews()
-        setupContainers()
         
-        let subviews: [UIView] = [
+        let children: [UIView] = [
             mapView,
-            informationStackView,
+            topBannerContainerView,
             floatingStackView,
             resumeButton,
             wayNameView,
-            bottomBannerContainerView,
-            instructionsBannerContentView
+            bottomBannerContainerView
         ]
         
-        subviews.forEach(addSubview(_:))
+        addSubviews(children)
     }
     
     open override func prepareForInterfaceBuilder() {
         super.prepareForInterfaceBuilder()
         DayStyle().apply()
-        [mapView, instructionsBannerView, lanesView, bottomBannerContainerView, nextBannerView].forEach( { $0.prepareForInterfaceBuilder() })
+        [mapView, topBannerContainerView, bottomBannerContainerView].forEach( { $0.prepareForInterfaceBuilder() })
         wayNameView.text = "Street Label"
     }
     
     private func updateDelegates() {
         mapView.navigationMapViewDelegate = delegate
         mapView.courseTrackingDelegate = delegate
-        instructionsBannerView.delegate = delegate
-        instructionsBannerView.instructionDelegate = delegate
-        nextBannerView.instructionDelegate = delegate
     }
 }