+++ MapboxNavigation/NavigationViewController.swift
@@ -2,6 +2,7 @@ import UIKit
 import MapboxCoreNavigation
 import MapboxDirections
 import MapboxSpeech
+import AVFoundation
 import Mapbox
 #if canImport(CarPlay)
 import CarPlay
@@ -22,7 +23,7 @@ public typealias ContainerViewController = UIViewController & NavigationComponen
  */
 
 @objc(MBNavigationViewController)
-open class NavigationViewController: UIViewController {
+open class NavigationViewController: UIViewController, NavigationStatusPresenter {
     
     /** 
      A `Route` object constructed by [MapboxDirections](https://mapbox.github.io/mapbox-navigation-ios/directions/).
@@ -137,14 +138,6 @@ open class NavigationViewController: UIViewController {
      */
     @objc public var shouldManageApplicationIdleTimer = true
     
-    /**
-     Bool which should be set to true if a CarPlayNavigationView is also being used.
-     */
-    @objc public var isUsedInConjunctionWithCarPlayWindow = false {
-        didSet {
-            mapViewController?.isUsedInConjunctionWithCarPlayWindow = isUsedInConjunctionWithCarPlayWindow
-        }
-    }
     
     var isConnectedToCarPlay: Bool {
         if #available(iOS 12.0, *) {
@@ -156,6 +149,8 @@ open class NavigationViewController: UIViewController {
     
     var mapViewController: RouteMapViewController?
     
+    var topViewController: ContainerViewController?
+    
     var bottomViewController: ContainerViewController?
     
     var navigationComponents: [NavigationComponent] {
@@ -163,6 +158,11 @@ open class NavigationViewController: UIViewController {
         if let mvc = mapViewController {
             components.append(mvc)
         }
+        
+        if let topViewController = topViewController {
+            components.append(topViewController)
+        }
+        
         if let bottomViewController = bottomViewController {
             components.append(bottomViewController)
         }
@@ -176,12 +176,7 @@ open class NavigationViewController: UIViewController {
     
     var styleManager: StyleManager!
     
-    var currentStatusBarStyle: UIStatusBarStyle = .default {
-        didSet {
-            mapViewController?.instructionsBannerView.backgroundColor = InstructionsBannerView.appearance().backgroundColor
-            mapViewController?.instructionsBannerContentView.backgroundColor = InstructionsBannerContentView.appearance().backgroundColor
-        }
-    }
+    var currentStatusBarStyle: UIStatusBarStyle = .default
     
     open override var preferredStatusBarStyle: UIStatusBarStyle {
         get {
@@ -216,6 +211,10 @@ open class NavigationViewController: UIViewController {
 
         NavigationSettings.shared.distanceUnit = route.routeOptions.locale.usesMetric ? .kilometer : .mile
         
+        styleManager = StyleManager()
+        styleManager.delegate = self
+        styleManager.styles = options?.styles ?? [DayStyle(), NightStyle()]
+        
         let bottomBanner = options?.bottomBanner ?? {
             let viewController = BottomBannerViewController()
             viewController.delegate = self
@@ -223,7 +222,17 @@ open class NavigationViewController: UIViewController {
         }()
         bottomViewController = bottomBanner
 
-        let mapViewController = RouteMapViewController(navigationService: self.navigationService, delegate: self, bottomBanner: bottomBanner)
+        if let customBanner = options?.topBanner {
+            topViewController = customBanner
+        } else {
+            let defaultBanner = TopBannerViewController(nibName: nil, bundle: nil)
+            defaultBanner.delegate = self
+            defaultBanner.statusView.addTarget(self, action: #selector(NavigationViewController.didChangeSpeed(_:)), for: .valueChanged)
+            topViewController = defaultBanner
+        }
+        
+        let mapViewController = RouteMapViewController(navigationService: self.navigationService, delegate: self, topBanner: topViewController!, bottomBanner: bottomBanner)
+        
         self.mapViewController = mapViewController
         mapViewController.destination = route.legs.last?.destination
         mapViewController.view.translatesAutoresizingMaskIntoConstraints = false
@@ -233,17 +242,16 @@ open class NavigationViewController: UIViewController {
             return map.view.constraintsForPinning(to: parent.view)
         }
         
-
+        //Manually update the map style since the RMVC missed the "map style change" notification when the style manager was set up.
+        if let currentStyle = styleManager.currentStyle {
+            updateMapStyle(currentStyle, animated: false)
+        }
         
         //Do not start the navigation session until after you create the MapViewController, otherwise you'll miss important messages.
         self.navigationService.start()
         
         mapViewController.view.pinInSuperview()
         mapViewController.reportButton.isHidden = !showsReportFeedback
-        
-        styleManager = StyleManager()
-        styleManager.delegate = self
-        styleManager.styles = options?.styles ?? [DayStyle(), NightStyle()]
         
         if !(route.routeOptions is NavigationRouteOptions) {
             print("`Route` was created using `RouteOptions` and not `NavigationRouteOptions`. Although not required, this may lead to a suboptimal navigation experience. Without `NavigationRouteOptions`, it is not guaranteed you will get congestion along the route line, better ETAs and ETA label color dependent on congestion.")
@@ -271,6 +279,12 @@ open class NavigationViewController: UIViewController {
         // This is optional and lazy so it can be mutated by the developer after init.
         _ = voiceController
         view.clipsToBounds = true
+
+ 
+        guard let firstInstruction = navigationService.routeProgress.currentLegProgress.currentStepProgress.currentVisualInstruction else {
+            return
+        }
+        navigationService(navigationService, didPassVisualInstructionPoint: firstInstruction, routeProgress: navigationService.routeProgress)
     }
     
     open override func viewWillAppear(_ animated: Bool) {
@@ -280,6 +294,7 @@ open class NavigationViewController: UIViewController {
             UIApplication.shared.isIdleTimerDisabled = true
         }
         
+        notifyUserAboutLowVolumeIfNeeded()
     }
     
     open override func viewWillDisappear(_ animated: Bool) {
@@ -289,6 +304,15 @@ open class NavigationViewController: UIViewController {
             UIApplication.shared.isIdleTimerDisabled = false
         }
         
+    }
+    
+    func notifyUserAboutLowVolumeIfNeeded() {
+        guard !(navigationService.locationManager is SimulatedLocationManager) else { return }
+        guard !NavigationSettings.shared.voiceMuted else { return }
+        guard AVAudioSession.sharedInstance().outputVolume <= NavigationViewMinimumVolumeForWarning else { return }
+        
+        let title = String.localizedStringWithFormat(NSLocalizedString("DEVICE_VOLUME_LOW", bundle: .mapboxNavigation, value: "%@ Volume Low", comment: "Format string for indicating the device volume is low; 1 = device model"), UIDevice.current.model)
+        showStatus(title: title, spinner: false, duration: 3, animated: true, interactive: false)
     }
     
     // MARK: Containerization
@@ -327,6 +351,12 @@ open class NavigationViewController: UIViewController {
         // This way, there is always just one notification.
         UIApplication.shared.applicationIconBadgeNumber = 1
         UIApplication.shared.applicationIconBadgeNumber = 0
+    }
+    
+    public func showStatus(title: String, spinner: Bool, duration: TimeInterval, animated: Bool, interactive: Bool) {
+        navigationComponents.compactMap({ $0 as? NavigationStatusPresenter }).forEach {
+            $0.showStatus(title: title, spinner: spinner, duration: duration, animated: animated, interactive: interactive)
+        }
     }
 }
 
@@ -392,6 +422,12 @@ extension NavigationViewController: RouteMapViewControllerDelegate {
     
     @objc public func label(_ label: InstructionLabel, willPresent instruction: VisualInstruction, as presented: NSAttributedString) -> NSAttributedString? {
         return delegate?.label?(label, willPresent: instruction, as: presented)
+    }
+    
+    @objc func mapViewController(_ mapViewController: RouteMapViewController, didCenterOn location: CLLocation) {
+        navigationComponents.compactMap({$0 as? NavigationMapInteractionObserver}).forEach {
+            $0.navigationViewController(didCenterOn: location)
+        }
     }
 }
 
@@ -573,8 +609,12 @@ extension NavigationViewController: StyleManagerDelegate {
     
     @objc(styleManager:didApplyStyle:)
     public func styleManager(_ styleManager: StyleManager, didApply style: Style) {
+        updateMapStyle(style)
+    }
+    
+    private func updateMapStyle(_ style: Style, animated: Bool = true) {
         if mapView?.styleURL != style.mapStyleURL {
-            mapView?.style?.transition = MGLTransition(duration: 0.5, delay: 0)
+            mapView?.style?.transition = MGLTransition(duration: animated ? 0.5 : 0, delay: 0)
             mapView?.styleURL = style.mapStyleURL
         }
         
@@ -584,6 +624,98 @@ extension NavigationViewController: StyleManagerDelegate {
     
     @objc public func styleManagerDidRefreshAppearance(_ styleManager: StyleManager) {
         mapView?.reloadStyle(self)
+    }
+}
+// MARK: - TopBannerViewController
+// MARK: Status View Actions
+extension NavigationViewController {
+    @objc func didChangeSpeed(_ statusView: StatusView) {
+        let displayValue = 1+min(Int(9 * statusView.value), 8)
+        statusView.showSimulationStatus(speed: displayValue)
+
+        if let locationManager = navigationService.locationManager as? SimulatedLocationManager {
+            locationManager.speedMultiplier = Double(displayValue)
+        }
+    }
+}
+// MARK: TopBannerViewControllerDelegate
+extension NavigationViewController: TopBannerViewControllerDelegate {    
+    public func topBanner(_ banner: TopBannerViewController, didSwipeInDirection direction: UISwipeGestureRecognizer.Direction) {
+        let progress = navigationService.routeProgress
+        let route = progress.route
+        
+        if direction == .down {
+            banner.displayStepsTable()
+            
+            
+            if banner.isDisplayingPreviewInstructions {
+                mapViewController?.recenter(self)
+            }
+            
+        } else if direction == .right {
+            // prevent swiping when step list is visible
+            if banner.isDisplayingSteps {
+                return
+            }
+            
+            guard let currentStepIndex = banner.currentPreviewStep?.1 else { return }
+            let remainingSteps = progress.remainingSteps
+            let prevStepIndex = currentStepIndex.advanced(by: -1)
+            guard prevStepIndex >= 0 else { return }
+            
+            let prevStep = remainingSteps[prevStepIndex]
+            preview(step: prevStep, in: banner, remaining: remainingSteps, route: route)
+        } else if direction == .left {
+            // prevent swiping when step list is visible
+            if banner.isDisplayingSteps {
+                return
+            }
+            
+            let remainingSteps = navigationService.router.routeProgress.remainingSteps
+            let currentStepIndex = banner.currentPreviewStep?.1
+            let nextStepIndex = currentStepIndex?.advanced(by: 1) ?? 0
+            guard nextStepIndex < remainingSteps.count else { return }
+            
+            let nextStep = remainingSteps[nextStepIndex]
+            preview(step: nextStep, in: banner, remaining: remainingSteps, route: route)
+        }
+    }
+    
+    public func preview(step: RouteStep, in banner: TopBannerViewController, remaining: [RouteStep], route: Route, animated: Bool = true) {
+        guard let leg = route.leg(containing: step) else { return }
+        guard let legIndex = route.legs.index(of: leg) else { return }
+        guard let stepIndex = leg.steps.index(of: step) else { return }
+        let nextStepIndex = stepIndex + 1
+        
+        let legProgress = RouteLegProgress(leg: leg, stepIndex: stepIndex)
+        guard let upcomingStep = legProgress.upcomingStep else { return }
+        
+        let previewBanner: CompletionHandler = {
+            banner.preview(step: legProgress.currentStep, maneuverStep: upcomingStep, distance: legProgress.currentStep.distance, steps: remaining)
+        }
+        
+        mapViewController?.center(on: upcomingStep, route: route, legIndex: legIndex, stepIndex: nextStepIndex, animated: animated, completion: previewBanner)
+        
+        
+        
+    }
+    
+    public func topBanner(_ banner: TopBannerViewController, didSelect legIndex: Int, stepIndex: Int, cell: StepTableViewCell) {
+        let progress = navigationService.routeProgress
+        let legProgress = RouteLegProgress(leg: progress.route.legs[legIndex], stepIndex: stepIndex)
+        let step = legProgress.currentStep
+        self.preview(step: step, in: banner, remaining: progress.remainingSteps, route: progress.route, animated: false)
+        banner.dismissStepsTable()
+    }
+    
+    public func topBanner(_ banner: TopBannerViewController, didDisplayStepsController: StepsViewController) {
+        mapViewController?.recenter(self)
+    }
+}
+
+fileprivate extension Route {
+    func leg(containing step: RouteStep) -> RouteLeg? {
+        return legs.first { $0.steps.contains(step) }
     }
 }
 
@@ -602,3 +734,17 @@ extension NavigationViewController: BottomBannerViewControllerDelegate {
     }
 }
 
+// MARK: - CarPlayConnectionObserver
+
+extension NavigationViewController: CarPlayConnectionObserver {
+    public func didConnectToCarPlay() {
+        navigationComponents.compactMap({$0 as? CarPlayConnectionObserver}).forEach {
+            $0.didConnectToCarPlay()
+        }
+    }
+    public func didDisconnectFromCarPlay() {
+        navigationComponents.compactMap({$0 as? CarPlayConnectionObserver}).forEach {
+            $0.didDisconnectFromCarPlay()
+        }
+    }
+}